added accelerators to property and text field. enter submits property and ctrl+enter submits method

+++ src/com/chris3000/p4ming/viewer/text/P4Text.java
@@ -110,7 +110,7 @@ public class P4Text {
 		if (showFramerate){
 			p.fill(255);
 			p.textFont(font, 14);
-			p.text(""+p.frameRate,5,p.height-30);
+			p.text(""+(int)(p.frameRate),5,p.height-30);
 		}
 	}
